requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
cm==0.1a2
crypto==1.4.1
cymysql==0.9.14
Django==2.2.5
django-cymysql==2.2.0
django-session-timeout==0.0.4
django-debug-toolbar==2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
Jinja2==2.10.3
MarkupSafe==1.1.1
Naked==0.1.31
openpyxl==3.0.3
paragraph==1.1.0
Pillow==6.2.1
pytz==2019.2
PyYAML==5.3
reportlab==3.5.34
requests==2.22.0
shellescape==3.4.1
sqlparse==0.3.0
urllib3==1.25.8
